+++ stimline_idex/__init__.py
@@ -1,4 +1,4 @@
-from .api.v1 import IDEXClient
-from .api.v1.auth import ApiKeyAuth, JwtAuth
+from .v1.api import IDEXClient
+from .v1.api.auth import ApiKeyAuth, JwtAuth
 
 __all__ = ["IDEXClient", "ApiKeyAuth", "JwtAuth"]
