+++ stimline_idex/v1/api/modules/customers.py
@@ -1,17 +1,16 @@
 import logging
 from typing import Any, Optional, Union, overload
 
-from ....data_schemas.v1.assets import Well
+from ....v1.data_schemas.assets import Customer
 from ..api import IDEXApi
 
 
-class Wells:
+class Customers:
     def __init__(self, api: IDEXApi) -> None:
         self._api = api
 
     @overload
-    def get(self, *, id: str) -> Well: ...
-
+    def get(self, *, id: str) -> Customer: ...
     @overload
     def get(
         self,
@@ -22,11 +21,10 @@ class Wells:
         skip: Optional[int] = None,
         order_by: Optional[str] = None,
         include_soft_delete: Optional[bool] = False,
-    ) -> list[Well]: ...
+    ) -> list[Customer]: ...
 
     def get(
         self,
-        *,
         id: Optional[str] = None,
         filter: Optional[str] = None,
         select: Optional[list[str]] = None,
@@ -34,14 +32,14 @@ class Wells:
         skip: Optional[int] = None,
         order_by: Optional[str] = None,
         include_soft_delete: Optional[bool] = False,
-    ) -> Union[Well, list[Well]]:
+    ) -> Union[Customer, list[Customer]]:
         """
-        Get `Well` object(s).
+        Get `Customer` object(s).
 
         Parameters
         ----------
         id : Optional[str]
-            Well to retrieve.
+            Customer to retrieve.
         filter : Optional[str]
             OData filter string.
         select : list[str] | None
@@ -57,15 +55,14 @@ class Wells:
 
         Returns
         -------
-        Union[Well, list[Well]]
-            The `Well` object(s).
+        Union[Customer, list[Customer]]
+            The `Customer` object(s).
 
         """
         if id is not None:
-            # Get singular well
-            logging.debug(f"Getting Well with ID: {id}")
-            data = self._api.get(url=f"Wells/{id}")
-            return Well.model_validate(data.json())
+            logging.debug(f"Getting Customer with ID: {id}")
+            data = self._api.get(url=f"Customers/{id}")
+            return Customer.model_validate(data.json())
 
         params: dict[str, Any] = {}
         if filter is not None:
@@ -80,17 +77,17 @@ class Wells:
         if order_by is not None:
             params["$orderby"] = order_by
 
-        data = self._api.get(url="Wells", params=params)
+        data = self._api.get(url="Customers", params=params)
 
         if data.status_code == 204:
             return []
 
-        wells = [Well.model_validate(row) for row in data.json()]
+        customers = [Customer.model_validate(row) for row in data.json()]
 
         if include_soft_delete:
-            return wells
+            return customers
 
-        return [well for well in wells if well.deleted_date is None]
+        return [customer for customer in customers if customer.deleted_date is None]
 
     def _check_select(self, select: list[str]) -> list[str]:
         important_fields = ["id"]
