+++ stimline_idex/v1/data_schemas/assets.py
