+++ stimline_idex/v1/data_schemas/base.py
