+++ stimline_idex/v1/data_schemas/events/soe.py
